Many compiler fixes, but there are linker errors likely due to missing 'virtual' delcarations

+++ src/verilog/ast/types/node.h
@@ -240,7 +240,7 @@ inline uint32_t Node::get_val() const {
 std::string Node::stringify() const {
     STRINGIFY_BEGIN(Node);
     STRINGIFY_BASE_VAL(common);
-    STRINGIFY_BASE_VAL(tag);
+    STRINGIFY_BASE_VAL((uint32_t)tag);
     STRINGIFY_END();
 }
